requirements.txt
streamlit>=1.30,<2.0
opencv-python>=4.8,<5.0
numpy>=1.24,<2.0
pdf2image>=1.16,<2.0
PyMuPDF>=1.23,<2.0
python-dotenv>=1.0,<2.0
sentence-transformers>=2.2,<4.0
torch>=2.0,<3.0
torchvision>=0.15,<1.0
faiss-cpu>=1.7,<2.0
pytesseract>=0.3.10,<1.0
Pillow>=10.0,<12.0
